backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.5",
    "asgi-lifespan>=2.1.0",
    "asyncpg>=0.30.0",
    "bcrypt==4.0.1",
    "fastapi>=0.116.1",
    "fastapi-limiter>=0.1.6",
    "greenlet>=3.2.4",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pyjwt[crypto]>=2.10.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.1.0",
    "pytest-order>=1.3.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "secure>=1.0.1",
    "sqlalchemy[asyncio]>=2.0.43",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
]
